Periscopai React Frontend (#15)
* Just a skeleton of React Application

* babel installation

* Some basic stuff based on the react tutorial I am going through. Sets up all the JS stuff requirement.
This little exercise should result in our mini-front for prototyping purposes

* some doc

* A very meaningless skeleton

* Added support for webpack to bundle the application.

* - integrated webpack to create the bundle.js
- replaced live-server with webpack server which cause the bundle to be generate in memory
- broke down the dummy app into separate component files
- very basic request made to the backend

* Documented process around firefox-dev

* Added a bit of styling with SCSS

* minor mods

* CSS Styling

* Very painful implementation of WebRTC.

Right now I am just discovering the API. It is actually not a webRTC implementation.
All I am able to do is to capture the video/audio from the browser.

* Added some documentation. Very minor.

* Fixed broken links

* fixed some typos

* Removed the model and moved the sequencer code into streaming (this made more sense to manage state).

Fine tuned the JSX so that the video tag and the image (when sequencer is off) is nicely overlapped.
Moved all the WebRTC code into the Streaming React component.

Still sucks a lot, but sucks less

* stopping the video source (browser request)